Remove the minimal version requirement
use Module::Build::Tiny 0.034;

Well, this version is too new to be shipped in most of distributions(I mean via package manager, users can install cpan from the command line, but when users want to use linode-cli, he must use package manager to install, and here comes the dependency).

Is it OK to remove the version so we can continue to build it on old system? 

Thanks.